Implement index locking to prevent phantom
