+++ src/main/java/org/vanilladb/core/storage/index/btree/BTreeIndex.java
@@ -231,8 +231,6 @@ public class BTreeIndex extends Index {
 			leaf.close();
 			leaf = null;
 		}
-		// release all locks on index structure
-		ccMgr.releaseIndexLocks();
 		dirsMayBeUpdated = null;
 	}
