+++ src/main/java/org/vanilladb/core/storage/tx/concurrency/ReadCommittedConcurrencyMgr.java
@@ -111,4 +111,16 @@ public class ReadCommittedConcurrencyMgr extends ConcurrencyMgr {
 		// releases IS lock to allow phantoms
 		lockTbl.release(dataFileName, txNum, LockTable.IS_LOCK);
 	}
+	
+	@Override
+	public void modifyLeafBlock(BlockId blk) {
+		lockTbl.xLock(blk, txNum);
+	}
+	
+	@Override
+	public void readLeafBlock(BlockId blk) {
+		lockTbl.sLock(blk, txNum);
+		// releases S lock to allow phantoms
+		lockTbl.release(blk, txNum, LockTable.S_LOCK);
+	}
 }
