+++ src/main/java/org/vanilladb/core/storage/tx/concurrency/SerializableConcurrencyMgr.java
@@ -91,4 +91,14 @@ public class SerializableConcurrencyMgr extends ConcurrencyMgr {
 	public void readIndex(String dataFileName) {
 		lockTbl.isLock(dataFileName, txNum);
 	}
+	
+	@Override
+	public void modifyLeafBlock(BlockId blk) {
+		lockTbl.xLock(blk, txNum);
+	}
+	
+	@Override
+	public void readLeafBlock(BlockId blk) {
+		lockTbl.sLock(blk, txNum);
+	}
 }
